Control.Monad.Supervisor

Description

A supervisor monad that explore the execution tree of an internal monad and define extra behaviours thanks to flexible instance definitions for each particular purpose. It can inject new behaviours for backtracking, trace generation, testing, transaction rollbacks etc The supervisor monad is used in the package MFlow to control the routing, tracing, state management, back button management and navigation in general

Documentation

data Control a

The internal computation can be reexecuted, proceed forward or backward

Constructors

Control a

Forward a

Backward

newtype Sup m a

The supervisor add a Control wrapper that is interpreted by the monad instance

Constructors

Sup

Fields

runSup :: m (Control a)

class MonadState s m => Supervise s m where

The supervise class add two general modifiers that can be applied:

Methods

supBack

Arguments

:: s

-> m ()
When the computation goes back, by default the previous state is restored This procedure can change that behaviour.

supervise

Arguments

:: m (Control a)

-> m (Control a)
When the conputation has been executed this method is an opportunity for modifying the result By default: supervise= id

breturn :: Monad m => a -> Sup m a

Flag the computation that executes breturn as a control point.

When the computation is going back, it will be re-executed (see the monad definition)